add the Votes struct
Now we have a struct Votes that can be built from a LuminanceImage.
It contains the votes (between 0 and 63) for each pixel (or None for no vote).
It reduces the size of the memory allocated for the votes (i32 -> u8) and makes it more clear if a position contains a vote or not.
It also reduces the need to cast at multiple places.